requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.3.0
attrs==23.2.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==42.0.7
dataclasses-json==0.6.5
distro==1.9.0
et-xmlfile==1.1.0
frozenlist==1.4.1
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
jsonpatch==1.33
jsonpointer==2.4
langchain==0.1.17
langchain-community==0.0.37
langchain-core==0.1.51
langchain-openai==0.1.6
langchain-text-splitters==0.0.1
langsmith==0.1.54
marshmallow==3.21.2
multidict==6.0.5
mypy-extensions==1.0.0
numpy==1.26.4
openai==1.26.0
openpyxl==3.1.2
orjson==3.10.3
packaging==23.2
pandas==2.2.2
pdfminer.six==20231228
pdfplumber==0.11.0
pillow==10.3.0
pycparser==2.22
pydantic==2.7.1
pydantic_core==2.18.2
PyPDF2==3.0.1
pypdfium2==4.29.0
python-dateutil==2.9.0.post0
python-decouple==3.8
pytz==2024.1
pywget==0.31
PyYAML==6.0.1
regex==2024.4.28
requests==2.31.0
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.30
tenacity==8.2.3
tiktoken==0.6.0
tqdm==4.66.4
typing-inspect==0.9.0
typing_extensions==4.11.0
tzdata==2024.1
urllib3==2.2.1
wget==3.2
yarl==1.9.4
